adding changes to get balance

See merge request campuses/cle/jan-2022/java/student-pairs/java-capstone-module-2-team-0!15

+++ capstone/tenmo-client/src/main/java/com/techelevator/tenmo/services/ConsoleService.java
@@ -90,7 +90,7 @@ public class ConsoleService {
 
     public void printBalance(){
         UserService userService = new UserService();
-        BigDecimal balance = userService.getBalance();
+        Integer balance = userService.getBalance();
         System.out.println("--------------------------------------------");
         System.out.println("Your balance is " + balance);
         System.out.println("--------------------------------------------");
